+++ changelog.md
@@ -971,6 +971,8 @@ Security in case of vulnerabilities.
 ### fixed
 - extract title from wysiwyg / correction
 ## [2.2.1] - 04/12/2023
+### Added
+- style depend of renderer "style-list.#renderer key#"
 ### Modified
 - display preview interface when error on template parsing